requirements.txt
# ----------------------
# Flask
# ----------------------
flask
pandas
numpy
seaborn
matplotlib
pickle-mixin
sklearn
wordcloud
joblib
Pillow


# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gevent
gunicorn
